% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_means_tables.R
\name{join_means_tables}
\alias{join_means_tables}
\title{join_means_tables}
\usage{
join_means_tables(means_data_earlier, means_data_later)
}
\arguments{
\item{means_data_earlier}{Dataframe from summarize_schools_by_loc}

\item{means_data_later}{Dataframe from summarize_schools_by_loc}
}
\description{
Sort table of summarized schools data.
}
\examples{
join_means_tables(df_a, df_b)
}
