% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_changes_table.R
\name{sort_changes_table}
\alias{sort_changes_table}
\title{sort_means_table}
\usage{
sort_changes_table(.data, sort_var)
}
\arguments{
\item{.data}{Dataframe from summarize_schools function with a statname.}

\item{sort_var}{Variable to sort by.}

\item{desc}{Descending}
}
\description{
Sort table of summarized schools data.
}
\examples{
means_df \%>\% sort_means_table(sort_var = desc(mean_ll_lr_ratio), name_var = statcode)
}
