% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_schools_by_loc.R
\name{summarize_schools_by_loc}
\alias{summarize_schools_by_loc}
\title{summarize_schools_by_loc}
\usage{
summarize_schools_by_loc(.data, loc)
}
\arguments{
\item{.data}{Dataframe with DUO school data, including \code{leerlingen}, \code{onderwijzers}, \code{gewicht_ll}, \code{ll_lr_ratio}, etc.}

\item{loc}{Column to summarize by.}
}
\description{
Summarize school data by a particular column.
}
\examples{
schools_df \%>\% summarize_schools_by_loc(loc = statcode)
}
